requirements.txt
Flask==3.0.3
python-dotenv==1.1.0
requests==2.32.3
